share Tower API CUD operations
