+++ app/models/manageiq/providers/ansible_tower/automation_manager/configuration_script_source.rb
@@ -2,4 +2,5 @@ class ManageIQ::Providers::AnsibleTower::AutomationManager::ConfigurationScriptS
   ManageIQ::Providers::ExternalAutomationManager::ConfigurationScriptSource
 
   include ManageIQ::Providers::AnsibleTower::Shared::AutomationManager::ConfigurationScriptSource
+  include ManageIQ::Providers::AnsibleTower::Shared::AutomationManager::TowerApi
 end
