+++ app/models/manageiq/providers/embedded_ansible/automation_manager/configuration_script_source.rb
@@ -1,5 +1,5 @@
-class ManageIQ::Providers::EmbeddedAnsible::AutomationManager::ConfigurationScriptSource <
-  ManageIQ::Providers::EmbeddedAutomationManager::ConfigurationScriptSource
+class ManageIQ::Providers::EmbeddedAnsible::AutomationManager::ConfigurationScriptSource < ManageIQ::Providers::EmbeddedAutomationManager::ConfigurationScriptSource
 
   include ManageIQ::Providers::AnsibleTower::Shared::AutomationManager::ConfigurationScriptSource
+  include ManageIQ::Providers::AnsibleTower::Shared::AutomationManager::TowerApi
 end
